move updated viz integration tests into new PR, to merge back in after isaac sim CI is moved to latest develop
